Extract bignum and time into modules
