+++ edgedb-protocol/src/lib.rs
@@ -1,5 +1,7 @@
 mod encoding;
 mod common;
+mod bignum;
+mod time;
 pub mod client_message;
 pub mod server_message;
 pub mod errors;
